build: tweak `FetchContent` usage to follow best practices
We should do a singular call to `FetchContent_MakeAvailable` as per the
documentation. Heed this advice which also allows us to reduce the scope
of the overridden variables. Once we migrate to a new enough CMake
minimum (3.25), we should be able to replace that with a `block` scope
to avoid the push/pop manual adjustment.