Skip testLinkFilelistWithDebugInfo off of macOS
Only the Apple-platforms support dsymutil and building/linking with the
swiftmodule, therefore, this test only applies to running on macOS where
we have dsymutil. I've also cleaned up the test a bit to make it a bit
clearer what the goals were.
We're trying to verify that the contents of the filelist passed into the
link job contains only the expected object file and does not include the
swiftmodule file.

Fixes: rdar://144378596