[Tests] Fix another early termination in tests
When running swift-driver test suite with an older compiler that doesn't
support bridging header chaining, the second part of the bridging header
caching test will just be a single job, from which the default executor
will run the swift-frontend in-place. This will cause the test suite to
early exit regardless of the state of the test suite.