Re-land [Planning] Avoid batching compile job twice
This re-land the change in 7ba9d485ab44cca5c61cf974679d6fba91feb1ca with
compatibility fixes and additional tests.

The fix makes sure the skipped jobs are returned in the planning as well for
compatibility reasons. Not after this change:

For the build systems that use swift-driver that doesn't support
incremental build (swiftpm), asking for an incremental planning will
still get the full list of jobs, but the skipped jobs might not be batched.
The planned jobs are still complete, but might not be optimally batched.

For the build systems that supports incremental build (swift-build), the
build planning is extracted from incremental state directly. But note
swift-driver lazily wrapped the compile jobs in Executor so swift-driver
can't return an empty list of jobs. Add a test to make sure that is the
case.