Append target variant triple on emit-module-job
When emitting the interfaces and modules for the target variant, we need
to pass the target-variant triple as the triple. This is only the case
for the emit-module-job though. The compile job emits zippered object
files, and there is only one compile job for a zippered module. Passing
the target variant triple as the last option to the frontend ensures
that the frontend uses the correct target triple when emitting
interfaces and binary modules for the target variant.

Fixes: rdar://146984980