FreeBSD: Fix rpath to swift runtimes
The build_os in build-script includes the platform version number from
the target triple. The runtimes are installed to a location that does
not include the version number, resulting in the swift-driver failing
to launch due to not finding libswiftCore. Fixing the rpath to avoid
requiring folks to manually specify the correct location through
environment variables.