[-debug-module-path] Prefer the moduleOutputInfo.output path if available.
While testing the LLDB support for enhanced module tracking I noticed
that the driver currently always uses a relative path for
-debug-module-path, which is an obvious problem when compiling in a
different directory, or if the swift module output is in a different
path than the object file output. This patch addresses this by
preferring moduleOutputInfo.output if available.

rdar://164524241

(Relanding with a relaxed test that doesn't encode the name of a temporary file)